remvoe all old snipes code & configuration related
