+++ PoGo.NecroBot.CLI/PushNotificationListener.cs
@@ -17,11 +17,7 @@ namespace PoGo.NecroBot.CLI
         {
             PushNotificationClient.SendNotification(session, "Error occured", errorEvent.Message);
         }
-
-        public static void HandleEvent(SnipePokemonFoundEvent ev, ISession session)
-        {
-        }
-
+       
         public static void HandleEvent(EncounteredEvent ev, ISession session)
         {
         }
