+++ PoGo.NecroBot.Logic/Event/SnipePokemonFoundEvent.cs
@@ -4,11 +4,9 @@ namespace PoGo.NecroBot.Logic.Event
 {
     public class SnipePokemonFoundEvent : IEvent
     {
-        public SniperInfo PokemonFound { get; set; }
-
         public override string ToString()
         {
-            return PokemonFound.Id.ToString();
+            return string.Empty;//remove this later
         }
     }
 }
