+++ PoGo.NecroBot.Logic/Tasks/CatchIncensePokemonsTask.cs
@@ -44,8 +44,7 @@ namespace PoGo.NecroBot.Logic.Tasks
                 if (session.Cache.Get(incensePokemon.EncounterId.ToString()) != null)
                     return; //pokemon been ignore before
 
-                if ((session.LogicSettings.UsePokemonSniperFilterOnly && !session.LogicSettings.PokemonToSnipe.Pokemon.Contains(pokemon.PokemonId))
-                    || (session.LogicSettings.UsePokemonToNotCatchFilter && session.LogicSettings.PokemonsNotToCatch.Contains(pokemon.PokemonId)))
+                if ((session.LogicSettings.UsePokemonToNotCatchFilter && session.LogicSettings.PokemonsNotToCatch.Contains(pokemon.PokemonId)))
                 {
                     Logger.Write(session.Translation.GetTranslation(TranslationString.PokemonIgnoreFilter,
                         session.Translation.GetPokemonTranslation(pokemon.PokemonId)));
